requirements.txt
pip>=23.3
wheel
omegaconf>=2.0.6
torch>=2.3.1
torchvision>=0.18.1
torchaudio>=2.3.1
faiss-cpu>=1.7.3
einops>=0.8.0
librosa>=0.10.2
praat-parselmouth
soundfile>=0.13.0
numpy>=1.25.2,<2.0.0
numba>=0.57.0
scipy>=1.15.0
requests>=2.32.3
aiohttp
ffmpy==0.3.1
ffmpeg-python>=0.2.0
matplotlib>=3.7.2
ipywidgets
IPython
beautifulsoup4
pycryptodome>=3.9.6,<4.0.0